chore: move story
