+++ packages/semi-ui/_base/_story/index.stories.jsx
@@ -1,7 +1,8 @@
 import React, { useMemo } from 'react';
-import { Button, Typography, Card, Tooltip, Tag, Avatar, Rating, Nav, Layout } from '../../index';
+import {Button, Typography, Card, Tooltip, Tag, Avatar, Rating, Nav, Layout, ConfigProvider, Select} from '../../index';
 import { IconHelpCircle, IconUser, IconStar, IconSetting } from '@douyinfe/semi-icons';
 import './index.scss';
+import semiGlobal from "../../_utils/semi-global";
 
 export default {
   title: 'Base',
@@ -113,4 +114,40 @@ export const TestAlwaysDarkLight = () => {
   }
 
   return <Demo />;
+};
+
+
+
+semiGlobal.config.overrideDefaultProps = {
+    Button: {
+        type: 'warning',
+    },
+    Select: {
+        zIndex: 2000,
+        getPopupContainer: () => document.querySelector('#popupContainer')
+    },
+    Tooltip: {
+        zIndex: 2001,
+        getPopupContainer: () => document.querySelector('#popupContainer'),
+        trigger:"click"
+    },
+};
+
+export const DefaultPropsDemo = () => {
+    return (
+        <div>
+            <ConfigProvider>
+                <div style={{ position: 'relative'}} id='popupContainer'></div>
+                <Tooltip content="zindex 2001">
+                    <Button>test</Button>
+                </Tooltip>
+                <Select
+                    optionList={[
+                        { value: 1, label: 1 },
+                        { value: 2, label: 2 },
+                    ]}
+                ></Select>
+            </ConfigProvider>
+        </div>
+    );
 };
