+++ packages/semi-ui/configProvider/_story/configProvider.stories.jsx
@@ -28,37 +28,3 @@ export const RTLFormDemo = () => (
         <RTLForm />
     </RTLWrapper>
 );
-
-semiGlobal.config.overrideDefaultProps = {
-    Button: {
-        type: 'warning',
-    },
-    Select: {
-        zIndex: 2000,
-        getPopupContainer: () => document.querySelector('#popupContainer')
-    },
-    Tooltip: {
-        zIndex: 2001,
-        getPopupContainer: () => document.querySelector('#popupContainer'),
-        trigger:"click"
-    },
-};
-
-export const DefaultPropsDemo = () => {
-    return (
-        <div>
-            <ConfigProvider>
-                <div style={{ position: 'relative'}} id='popupContainer'></div>
-                <Tooltip content="zindex 2001">
-                    <Button>test</Button>
-                </Tooltip>
-                <Select
-                    optionList={[
-                        { value: 1, label: 1 },
-                        { value: 2, label: 2 },
-                    ]}
-                ></Select>
-            </ConfigProvider>
-        </div>
-    );
-};
